# web/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CloudCertify API",
    "description": "CloudCertify API — a cloud certification quiz and exam prep platform.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://0.0.0.0:8080"
    },
    {
      "url": "https://api-cloudcertify.snowye.dev",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/quiz/{quizId}": {
      "get": {
        "tags": ["Quiz"],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDto"
                }
              }
            }
          }
        }
      }
    },
    "/quiz/{quizId}/start": {
      "post": {
        "tags": ["Quiz"],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartQuizRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/quiz/{quizId}/answer": {
      "post": {
        "tags": ["Quiz"],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerQuestionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/quiz/{quizId}/submit": {
      "post": {
        "tags": ["Quiz"],
        "parameters": [
          {
            "name": "quizId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitQuizResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitQuizResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitQuizResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AnswerQuestionRequestDto": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "integer",
            "format": "int32"
          },
          "questionId": {
            "type": "integer",
            "format": "int32"
          },
          "answerIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "DomainResult": {
        "required": ["domain", "correct", "total", "weight"],
        "type": "object",
        "properties": {
          "domain": {
            "type": "string"
          },
          "correct": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "QuestionDifficulty": {
        "enum": ["easy", "medium", "hard"],
        "type": "string"
      },
      "QuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "selectCount": {
            "type": "integer",
            "format": "int32"
          },
          "difficulty": {
            "$ref": "#/components/schemas/QuestionDifficulty"
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          }
        }
      },
      "QuestionType": {
        "enum": ["multiple_choice", "multiple_response"],
        "type": "string"
      },
...
      "QuizDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "submissionId": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDto"
            }
          }
        }
      },
      "QuizDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "iconName": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "quizProvider": {
            "$ref": "#/components/schemas/QuizProvider"
          },
          "quizLevel": {
            "$ref": "#/components/schemas/QuizLevel"
          },
          "slug": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "questionCount": {
            "type": "integer",
            "format": "int32"
          },
          "minQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "maxQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "subQuizzes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubquizDto"
            },
            "nullable": true
          }
        }
      },
      "QuizLevel": {
        "enum": ["foundational", "associate", "professional", "specialist"],
        "type": "string"
      },
      "QuizProvider": {
        "enum": ["azure", "aws", "gcp"],
        "type": "string"
      },
      "QuizResultAnswerDto": {
        "required": ["id", "isCorrect", "wasSelected"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isCorrect": {
            "type": "boolean"
          },
          "wasSelected": {
            "type": "boolean"
          }
        }
      },
      "QuizResultQuestionDto": {
        "required": ["id", "text", "type", "answers"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "concepts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "serviceCategory": {
            "type": "string",
            "nullable": true
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "explanation": {
            "type": "string",
            "nullable": true
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizResultAnswerDto"
            }
          }
        }
      },
      "StartQuizRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SubmitQuizRequestDto": {
        "type": "object",
        "properties": {
          "submissionId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SubmitQuizResponseDto": {
        "required": [
          "score",
          "totalQuestions",
          "correctCount",
          "scaledScore",
          "passed",
          "domainBreakdown",
          "questions"
        ],
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "totalQuestions": {
            "type": "integer",
            "format": "int32"
          },
          "correctCount": {
            "type": "integer",
            "format": "int32"
          },
          "scaledScore": {
            "type": "integer",
            "format": "int32"
          },
          "passed": {
            "type": "boolean"
          },
          "domainBreakdown": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DomainResult"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuizResultQuestionDto"
            }
          }
        }
      },
...
      "SubquizDto": {
        "required": ["id", "title", "domain", "slug", "isAvailable"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}